+++ Lab/Core/Src/global.c
@@ -8,4 +8,4 @@
 
 MODE mode = INIT;
 STATE line_1 = RED, line_2 = GREEN;
-uint16_t state_ID = 0;
+uint32_t state_ID = 0;
